% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asap.R
\name{asap}
\alias{asap}
\title{asap}
\usage{
asap(pca_input, as_file, sources = NULL, admixed = NULL)
}
\arguments{
\item{pca_input}{R data.frame of PCA with N PCs}

\item{as_file}{R data.frame with two columns: POP and A/S, POP column lists all populations to be considered, A/S indicatea whether the population should be considered as Admixed ('A') or as Source ('S')}

\item{sources}{R vector indicating the groups that should be considered as Sources}

\item{admixed}{R vector indicating the groups that should be considered as Admixed}
}
\value{
Returns a list containing the ancestries proportions per each Admixed group
}
\description{
asap function takes a PCA as input and analyses it via NNLS to describe admixed individuals as a mixture of sources groups.
}
\examples{
pca = read.table('data/TOY.pca') # OR
pca = read_eigen('data/TOY.pca.evec') # OR
pca = read_flash('data/TOY_flash.pca')

example_AS = read.table('data/Example_AS', header = T)

asap(pca_input = pca, as_file = example_AS) #OR
asap(pca_input = pca, sources = c('Source1','Source2','Source3'), admixed = c('Admixed1','Admixed2')
}
