% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{calculate_freq_metrics}
\alias{calculate_freq_metrics}
\title{Calculate Frequency Metrics}
\usage{
calculate_freq_metrics(
  freq_track,
  freq_range,
  start_time,
  end_time,
  wave,
  songspec,
  wl,
  wn,
  fsmooth,
  threshold,
  ovlp,
  fast = TRUE
)
}
\description{
Calculate Frequency Metrics
}
\keyword{internal}
