% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utillties.R
\name{check_pkg}
\alias{check_pkg}
\title{Check if a package is available}
\usage{
check_pkg(pkg_name)
}
\arguments{
\item{pkg_name}{Character name of the package}
}
\value{
TRUE if package is available, FALSE otherwise
}
\description{
Check if a package is available
}
\keyword{internal}
