% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{create_sliding_window}
\alias{create_sliding_window}
\title{Create Sliding Windows for Segment}
\usage{
create_sliding_window(i, x, window_size = 0.1, step_size = 0.005, ...)
}
\arguments{
\item{i}{Rendition number}

\item{x}{Data frame with segment information}

\item{window_size}{Window size in seconds}

\item{step_size}{Step size in seconds}

\item{...}{Additional arguments}
}
\value{
Data frame containing sliding window information
}
\description{
Internal function to generate overlapping time windows for a segment.
}
\keyword{internal}
