% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise.R
\name{.inverse_stft}
\alias{.inverse_stft}
\title{Inverse Short-Time Fourier Transform (overlap-add)}
\usage{
.inverse_stft(stft, wl, hop, win, original_length)
}
\description{
Inverse Short-Time Fourier Transform (overlap-add)
}
\keyword{internal}
