% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utillties.R
\name{ensure_pkgs}
\alias{ensure_pkgs}
\title{Ensure required packages are available, auto-installing if needed}
\usage{
ensure_pkgs(...)
}
\arguments{
\item{...}{Character names of required packages}
}
\value{
Nothing, called for side effects
}
\description{
Ensure required packages are available, auto-installing if needed
}
\keyword{internal}
