% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrogramming.R
\name{extract_single_spec}
\alias{extract_single_spec}
\title{Extract spectrogram for a single audio segment}
\usage{
extract_single_spec(x, wav_dir, wl, ovlp, wn, freq_range, ...)
}
\description{
Extract spectrogram for a single audio segment
}
\keyword{internal}
