% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{fast_pca}
\alias{fast_pca}
\title{Fast PCA Using IRLBA}
\usage{
fast_pca(matrix, n_components = 50, center = TRUE, scale = TRUE)
}
\arguments{
\item{matrix}{Input matrix}

\item{n_components}{Number of components}

\item{center}{Whether to center data}

\item{scale}{Whether to scale data}
}
\description{
Internal function for fast PCA using truncated SVD.
}
\keyword{internal}
