% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitch.R
\name{FF}
\alias{FF}
\title{Fundamental Frequency Estimation}
\usage{
FF(
  segment_row,
  wav_dir = NULL,
  method = c("cepstrum", "yin"),
  wl = 512,
  ovlp = 80,
  fmax = 1400,
  threshold = 10,
  plot = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{segment_row}{A single-row data frame containing segment information}

\item{wav_dir}{Directory containing wav files. If NULL, attempts to use an attribute from the input}

\item{method}{Pitch estimation method. Currently supports "cepstrum"}

\item{wl}{Window length for spectral analysis (default: 512)}

\item{ovlp}{Overlap percentage between windows (default: 80)}

\item{fmax}{Maximum frequency to consider (default: 1400 Hz)}

\item{threshold}{Threshold Amplitude threshold for cepstral method in \% (default = 10).}

\item{plot}{Logical, whether to plot the pitch estimation (default: FALSE)}
}
\value{
A matrix with two columns:
\itemize{
\item First column: Time points
\item Second column: Fundamental frequency values
}
}
\description{
Estimates the fundamental frequency (F0) of an audio segmentusing either cepstral analysis or YIN method
(currently under development) for a specified audio segment.
}
\examples{
\dontrun{
# Assuming 'segment' is a data frame with start_time and end_time columns
segment_row <- segment[1, ]
F0 <- FF(segment_row, wav_dir = "/path/to/wav/files")

# Plot the fundamental frequency
FF(segment_row, plot = TRUE)
}

}
\seealso{
\code{\link[seewave]{fund}} for underlying pitch detection method
}
