% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitch.R
\name{FF_single_row}
\alias{FF_single_row}
\title{Internal function to calculate the fundamental frequency of time series data derived from a single audio segment}
\usage{
FF_single_row(
  segment_row,
  wav_dir = NULL,
  method = c("cepstrum", "yin"),
  wl = 512,
  ovlp = 50,
  fmax = 1400,
  threshold = 10,
  plot = FALSE,
  ...
)
}
\arguments{
\item{segment_row}{A single-row data frame containing segment information}

\item{wav_dir}{Directory containing wav files.}

\item{method}{Pitch estimation method. Either "cepstrum" or "yin" (default: "cepstrum")}

\item{wl}{Window length for spectral analysis (default: 512)}

\item{ovlp}{Overlap percentage between windows (default: 80)}

\item{fmax}{Maximum frequency to consider (default: 1400 Hz)}

\item{threshold}{Threshold Amplitude threshold for cepstral method in \% (default = 10).}

\item{plot}{Logical, whether to plot the pitch estimation (default: FALSE)}
}
\value{
A matrix with two columns:
\itemize{
\item First column: Time points
\item Second column: Fundamental frequency values
}
}
\description{
Estimates the fundamental frequency (F0) of an audio segment using either cepstral analysis
or YIN method for a specified audio segment.
}
\seealso{
\code{\link[seewave]{fund}} for cepstral analysis method
\code{\link[librosa]{yin}} for YIN algorithm implementation
}
\keyword{internal}
