% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitch.R
\name{filter_short_segments}
\alias{filter_short_segments}
\title{Filter Short Segments from Template}
\usage{
filter_short_segments(template, min_samples)
}
\arguments{
\item{template}{Logical vector representing the template}

\item{min_samples}{Minimum number of samples for a valid segment}
}
\value{
Logical vector with short segments removed
}
\description{
Filter Short Segments from Template
}
\keyword{internal}
