% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bout.R
\name{find_bout}
\alias{find_bout}
\title{Detect Song Bouts in Audio Recordings}
\usage{
find_bout(x, ...)
}
\arguments{
\item{x}{An object to analyze, either a character path to WAV file
or a SAP object}

\item{...}{Additional arguments passed to specific methods}
}
\value{
A data frame with bout information or updated SAP object
}
\description{
A generic function to detect and analyze song bouts using RMS amplitude
thresholding with bandpass filtering.
}
\details{
This generic function supports bout detection through two methods:
\itemize{
\item Default method for individual WAV files
\item SAP object method for organized song recordings
}
}
\examples{
\dontrun{
# Detect bouts in single file
bouts <- find_bout("path/to/song.wav")

# Detect bouts in SAP object
sap_obj <- find_bout(sap_object,
                     segment_type = "motifs",
                     summary = TRUE)
}

}
