% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{find_continuous_regions}
\alias{find_continuous_regions}
\title{Find Continuous Regions}
\usage{
find_continuous_regions(signal)
}
\arguments{
\item{signal}{Binary signal vector}
}
\value{
Matrix of start and end indices
}
\description{
Internal function to identify continuous regions in a binary signal.
}
\details{
Processes binary signal to:
\itemize{
\item Find transitions between states
\item Identify continuous regions
\item Handle edge cases
}
}
\keyword{internal}
