% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{find_segment_clusters}
\alias{find_segment_clusters}
\title{Find Clusters Using Modularity Optimization}
\usage{
find_segment_clusters(
  neighbor_graphs,
  modularity.fxn = 1,
  resolution = 0.8,
  algorithm = 1,
  n.start = 10,
  n.iter = 10,
  random.seed = 0,
  group.singletons = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{neighbor_graphs}{List of neighbor graphs}

\item{modularity.fxn}{Modularity function type}

\item{resolution}{Resolution parameter}

\item{algorithm}{Clustering algorithm}

\item{n.start}{Number of random starts}

\item{n.iter}{Maximum iterations}

\item{random.seed}{Random seed}

\item{group.singletons}{Whether to group singleton clusters}

\item{verbose}{Print progress messages}
}
\value{
Data frame with cluster assignments
}
\description{
Internal function to perform cluster detection.
}
\keyword{internal}
