% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{find_segment_neighbors}
\alias{find_segment_neighbors}
\title{Find Nearest Neighbors for Segments}
\usage{
find_segment_neighbors(
  segments_data,
  k.param = 20,
  prune.SNN = 1/15,
  compute.SNN = TRUE,
  n.pcs = 10
)
}
\arguments{
\item{segments_data}{Matrix or data frame of segment features}

\item{k.param}{Number of nearest neighbors}

\item{prune.SNN}{SNN pruning threshold}

\item{compute.SNN}{Whether to compute SNN graph}

\item{n.pcs}{Number of PCs to use}
}
\value{
List containing neighbor graphs
}
\description{
Internal function to compute nearest neighbors and SNN graph.
}
\keyword{internal}
