% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{frd_wrblr_int}
\alias{frd_wrblr_int}
\title{Calculate Frequency Range Detection (from warbleR)}
\usage{
frd_wrblr_int(
  wave,
  wl = 512,
  fsmooth = 0.1,
  threshold = 10,
  wn = "hanning",
  bp = NULL,
  ovlp = 50,
  dB.threshold = NULL
)
}
\description{
Calculate Frequency Range Detection (from warbleR)
}
\keyword{internal}
