% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utillties.R
\name{get_wav_indices}
\alias{get_wav_indices}
\title{Get Indices of WAV Files in SAP Object Metadata}
\usage{
get_wav_indices(sap, wav_files)
}
\arguments{
\item{sap}{A SAP object}

\item{wav_files}{A character vector of WAV file names}
}
\value{
A numeric vector representing the row indices of the specified WAV files
in the SAP object's metadata.
}
\description{
Finds the row indices of specified WAV files within the metadata of a Sound Analysis Pro (SAP) object.
}
