% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{GroupSingletons}
\alias{GroupSingletons}
\title{Group Singleton Clusters}
\usage{
GroupSingletons(ids, SNN, group.singletons = TRUE, verbose = TRUE)
}
\description{
Internal function to handle singleton clusters.
}
\keyword{internal}
