% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{h5_get_or_create_group}
\alias{h5_get_or_create_group}
\title{Get or Create an HDF5 Group}
\usage{
h5_get_or_create_group(parent, group_name)
}
\arguments{
\item{parent}{An \code{H5File} or \code{H5Group} object.}

\item{group_name}{Character name of the group to retrieve or create.}
}
\value{
The requested HDF5 group object.
}
\description{
Returns an existing group from an HDF5 parent object, or creates it if it
does not already exist.
}
\keyword{internal}
