% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{merge_similar_clusters}
\alias{merge_similar_clusters}
\title{Merge Similar Clusters Based on UMAP Proximity}
\usage{
merge_similar_clusters(data, umap_threshold = NULL)
}
\arguments{
\item{data}{Data frame containing cluster information}

\item{umap_threshold}{Distance threshold for merging clusters}
}
\value{
Data frame with merged cluster assignments
}
\description{
Internal function to merge clusters that are close in UMAP space.
}
\keyword{internal}
