% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npx.R
\name{new_npx}
\alias{new_npx}
\title{Internal Constructor for NPX Object}
\usage{
new_npx(
  metadata = data.frame(),
  base_path = character(),
  templates = new_template_collection(),
  detections = list(),
  misc = list(),
  version = "1.0.0"
)
}
\arguments{
\item{metadata}{Data frame with file metadata}

\item{base_path}{Character string of base directory}

\item{templates}{Template collection object}

\item{detections}{List of detection results}

\item{misc}{List of miscellaneous info}

\item{version}{Version string}
}
\value{
An NPX object of class "Npx"
}
\description{
Internal Constructor for NPX Object
}
\keyword{internal}
