% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{new_template_collection}
\alias{new_template_collection}
\title{Internal Constructor for Template Collection}
\usage{
new_template_collection()
}
\value{
A template_collection object with an empty but structured template collection
}
\description{
An internal constructor function for creating a new template collection
with a standardized structure for storing and managing sound analysis templates.
}
\details{
Creates a template collection object with three key components:
\itemize{
\item template_info: A data frame containing metadata about templates
\item template_list: A list to store corTemplate objects
\item template_matches: A list to store template detection results
}

The template_info data frame includes the following columns:
\itemize{
\item template_name: Name of the template
\item start_time: Start time of the template
\item end_time: End time of the template
\item duration: Duration of the template
\item freq_min: Minimum frequency
\item freq_max: Maximum frequency
\item threshold: Detection threshold
\item clip_name: Name of the audio clip
\item clip_path: Path to the audio clip
\item source_file: Original source file name
\item source_file_path: Path to the source file
\item creation_date: Date and time of template creation
}
}
\note{
This is an internal function not intended to be called directly by users.
It is used within the package for creating template collections.
}
\keyword{internal}
