% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNLS2_step.R
\name{nnls.mat2}
\alias{nnls.mat2}
\title{nnls.mat2}
\usage{
nnls.mat2(donors, recipients)
}
\arguments{
\item{donors}{Matrix with reference groups}

\item{recipients}{Matrix with target groups}
}
\value{
Returns matrix describing the admixed groups as a mixture of the source groups, along with the residuals
}
\description{
nnls.mat2 function solves nonnegative least squares problems. It requires two matrices, one ('donors') refers to the source groups, the second ('recipients') refers to the admixed groups
}
\examples{
\dontrun{
nnls.mat2(donors = my_source_individuals_matrix,recipients = my_admixed_individuals_matrix)
}
}
