% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope.R
\name{pad_amp_env}
\alias{pad_amp_env}
\title{Pad Amplitude Envelope for Alignment}
\usage{
pad_amp_env(
  segment_row,
  wav_dir = NULL,
  msmooth = NULL,
  max_pre_window = NULL,
  max_post_window = NULL,
  samples_per_second = NULL,
  pad_value = 0,
  ...
)
}
\arguments{
\item{segment_row}{Single row of segment data}

\item{wav_dir}{Directory containing WAV files}

\item{msmooth}{Smoothing parameters}

\item{max_pre_window}{Maximum pre-window time}

\item{max_post_window}{Maximum post-window time}

\item{samples_per_second}{Sampling rate}

\item{pad_value}{Value to use for padding}

\item{...}{Additional arguments}
}
\value{
Padded and aligned amplitude envelope vector
}
\description{
Internal function to pad and align amplitude envelopes for bout analysis.
}
\keyword{internal}
