% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{pad_array}
\alias{pad_array}
\title{Pad an array with zeros to match the target length}
\usage{
pad_array(arr, target_length)
}
\arguments{
\item{arr}{Array to pad}

\item{target_length}{Target length after padding}
}
\value{
Padded array
}
\description{
Pad an array with zeros to match the target length
}
\keyword{internal}
