% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utillties.R
\name{parallel_apply}
\alias{parallel_apply}
\title{parallel processing function}
\usage{
parallel_apply(indices, FUN, cores, use_preschedule = FALSE)
}
\description{
parallel processing function
}
\keyword{internal}
