% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{parallel_pca}
\alias{parallel_pca}
\title{Parallel PCA Implementation}
\usage{
parallel_pca(x, n_components = 50, n_cores = NULL, scale = TRUE)
}
\arguments{
\item{x}{Input matrix}

\item{n_components}{Number of components}

\item{n_cores}{Number of cores}

\item{scale}{Whether to scale data}
}
\description{
Internal function for parallel PCA computation.
}
\keyword{internal}
