% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitch.R
\name{pitch_goodness}
\alias{pitch_goodness}
\title{Internal function for processing single-row fundamental frequency estimation}
\usage{
pitch_goodness(
  segment_row,
  wav_dir = NULL,
  wl = 512,
  ovlp = 80,
  fmax = 1500,
  plot = TRUE,
  ...
)
}
\description{
Internal function for processing single-row fundamental frequency estimation
}
\keyword{internal}
