% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope.R
\name{plot_heatmap.matrix}
\alias{plot_heatmap.matrix}
\title{Plot Heatmap from Amplitude Matrix}
\usage{
\method{plot_heatmap}{matrix}(
  x,
  labels = NULL,
  color_palette = NULL,
  n_colors = 500,
  contrast = 3,
  main = "Amplitude Envelope Heatmap",
  ...
)
}
\arguments{
\item{x}{Matrix of amplitude envelopes with time_window attribute}

\item{labels}{Optional vector of labels to subset the matrix}

\item{color_palette}{Function generating color palette}

\item{n_colors}{Number of colors in heatmap}

\item{contrast}{Contrast factor for visualization}

\item{main}{Title for the heatmap}

\item{...}{Additional arguments}
}
\value{
A lattice plot object containing the heatmap visualization
}
\description{
Creates a heatmap visualization from a pre-computed amplitude envelope matrix.
}
\details{
Creates a heatmap visualization with the following features:
\itemize{
\item Label-based organization
\item Customizable color scheme
\item Automatic time scaling
\item Visual separation between labels
}
}
\examples{
\dontrun{
# Plot full matrix
plot_heatmap(sap_obj$features$motif$amp_env)

# Plot subset of labels
plot_heatmap(sap_obj$features$motif$amp_env,
             labels = c("a", "b"))
}

}
