% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npx.R
\name{print.Npx}
\alias{print.Npx}
\title{Print method for NPX objects}
\usage{
\method{print}{Npx}(x, ...)
}
\arguments{
\item{x}{An NPX object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for NPX objects
}
