% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{print.segment}
\alias{print.segment}
\title{Print method for segment objects}
\usage{
\method{print}{segment}(x, ...)
}
\arguments{
\item{x}{A segment object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for segment objects
}
