% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_resampling.R
\name{read.resampling}
\alias{read.resampling}
\title{read.resampling}
\usage{
read.resampling(path_tofiles, file_pattern, as_file, eigentype)
}
\arguments{
\item{path_tofiles}{string containing directory path to the resampling files.}

\item{file_pattern}{string containing common pattern to find the all PCAs obtained by resampling}

\item{as_file}{R data.frame with two columns: POP and A/S, POP column lists all populations to be considered, A/S indicates whether the population should be considered as Admixed ('A') or as Source ('S')}

\item{eigentype}{if present, PCA will be read through read_eigen() function, if absent PCA will be loaded via read_flash()}
}
\value{
Returns a table containing the ASAP results per each resampled set.
}
\description{
read.resampling reads all resampled PCAs and performs ASAP on each one. It returns a list containing all ASAP results per each resampling. The output can then be used in the se.resampling function, to estimate the standard error.
}
\examples{
\dontrun{
example_as = read.table('data/Example_AS', header=TRUE)
pca_jackknife = read.resampling(path_tofiles = 'data/', file_pattern = '*_Jack*', as_file = example_as, eigentype) #OR
pca_jackknife = read.resampling(path_tofiles = 'data/', file_pattern = '*_Jack*', as_file = example_as)
}
}
