% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_PCA.R
\name{read_eigen}
\alias{read_eigen}
\title{read_eigen}
\usage{
read_eigen(pca_input)
}
\arguments{
\item{pca_input}{points to the directory and pca.evec file}
}
\value{
Returns a PCA matrix with new header: POP ID PC1 PCN CC
}
\description{
read_eigen reads smartpca output from EIGENSTRAT, where the first column contains Family ID, the second sample ID, and the following the PCs
}
\examples{
\dontrun{
read_eigen(pca_input = 'data/TOY.pca.evec')
}
}
