% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{run_pca.Sap}
\alias{run_pca.Sap}
\title{Run PCA on SAP Object}
\usage{
\method{run_pca}{Sap}(
  x,
  segment_type = c("motifs", "syllables", "bouts", "segments"),
  data_type = "traj_mat",
  method = c("irlba", "base", "parallel"),
  n_components = 50,
  n_cores = NULL,
  diagnostic_plots = TRUE,
  scale_scores = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{A SAP object}

\item{segment_type}{Type of segments to analyze}

\item{data_type}{Type of data to analyze}

\item{method}{PCA method to use}

\item{n_components}{Number of components}

\item{n_cores}{Number of cores}

\item{diagnostic_plots}{Whether to show diagnostics}

\item{scale_scores}{Whether to scale scores}

\item{verbose}{Whether to print progress}

\item{...}{Additional arguments}
}
\value{
Updated SAP object with PCA results stored in features slot
}
\description{
Performs PCA on trajectory matrices stored in a SAP object.
}
\details{
Performs PCA with the following features:
\itemize{
\item Support for different segment types
\item Multiple PCA methods
\item Diagnostic visualizations
\item Results storage in SAP object
}
}
