% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{run_umap}
\alias{run_umap}
\title{Run UMAP Dimensionality Reduction}
\usage{
run_umap(x, ...)
}
\arguments{
\item{x}{An object to analyze, either a data frame with feature data
or a SAP object}

\item{...}{Additional arguments passed to specific methods}
}
\value{
UMAP coordinates or updated SAP object
}
\description{
A generic function to perform UMAP dimensionality reduction on feature data.
}
\details{
This generic function supports UMAP analysis through two methods:
\itemize{
\item Default method for feature data frames
\item SAP object method for organized song features
}
}
\examples{
\dontrun{
# Run UMAP on feature data frame
coords <- run_umap(features,
                   metadata_cols = c(1:5),
                   n_neighbors = 15)

# Run UMAP on SAP object
sap_obj <- run_umap(sap_object,
                    segment_type = "motifs",
                    data_type = "spectral_feature")
}

}
