% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{RunModularityClustering}
\alias{RunModularityClustering}
\title{Run Modularity Clustering}
\usage{
RunModularityClustering(
  SNN = matrix(),
  modularity = 1,
  resolution = 0.8,
  algorithm = 1,
  n.start = 10,
  n.iter = 10,
  random.seed = 0,
  print.output = TRUE,
  temp.file.location = NULL,
  edge.file.name = NULL
)
}
\description{
Internal function to run modularity optimization.
}
\keyword{internal}
