% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrogramming.R
\name{sample_rows}
\alias{sample_rows}
\title{Sample Rows from a Vector}
\usage{
sample_rows(rows, n)
}
\arguments{
\item{rows}{A vector of row indices}

\item{n}{Number of rows to sample}
}
\value{
A vector of sampled row indices
}
\description{
An internal helper function to sample a specified number of rows
from a given vector of row indices.
}
\keyword{internal}
