% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_resampling.R
\name{se.resampling}
\alias{se.resampling}
\title{se.resampling}
\usage{
se.resampling(nnls_main, nnls_resampling, chromovec)
}
\arguments{
\item{nnls_main}{list obtained from the asap() function on the main set.}

\item{nnls_resampling}{list contining multilpe asap() function results, on the resampled set.}

\item{chromovec}{a numeric vector containing the number of SNPs per each chromosome (ie. chromovec = rep(1000,times = 22))}
}
\value{
Returns a table containing the standard error per each target given the source group.
}
\description{
se.resampling estimates the Standard Error comparing a main ASAP run and multiple other ASAP runs, for example runs obtained with jackknife resampling. se.resampling takes three inputs: a list with the main ASAP result, a list with the resampling ASAP results, and a numeric vector containing the number of SNPs per each chromosome.
}
\examples{
\dontrun{
se.resampling(nnls_main = ASAP_main, nnls_resampling = ASAP_resampling, chromovec)
}
}
