% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment}
\alias{segment}
\title{Segment Audio into Syllables}
\usage{
segment(x, ...)
}
\arguments{
\item{x}{An object to segment, either a character path to WAV file
or a SAP object}

\item{...}{Additional arguments passed to specific methods}
}
\value{
A data frame with syllable information or updated SAP object
}
\description{
A generic function to segment audio recordings into syllables using
dynamic thresholding of spectrograms.
}
\details{
This generic function supports syllable segmentation through two methods:
\itemize{
\item Default method for individual WAV files
\item SAP object method for batch processing multiple recordings
}
}
\examples{
\dontrun{
# Segment single audio file
syllables <- segment("path/to/song.wav")

# Segment recordings in SAP object
sap_obj <- segment(sap_object,
                   segment_type = "bouts")
}

}
