% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{specprop_wrblr_int}
\alias{specprop_wrblr_int}
\title{Calculate Spectral Properties (from warbleR)}
\usage{
specprop_wrblr_int(spec, f = NULL, flim = NULL, ...)
}
\description{
Calculate Spectral Properties (from warbleR)
}
\keyword{internal}
