% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitch.R
\name{split_at_dips}
\alias{split_at_dips}
\title{Split Template at Local Minima in Goodness}
\usage{
split_at_dips(template, goodness, min_duration_samples)
}
\arguments{
\item{template}{Logical vector representing the template}

\item{goodness}{Numeric vector of goodness values}

\item{min_duration_samples}{Minimum number of samples for a valid segment}
}
\value{
Logical vector with segments split at significant dips
}
\description{
Split Template at Local Minima in Goodness
}
\keyword{internal}
