% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npx.R
\name{summary.Npx}
\alias{summary.Npx}
\title{Summary method for NPX objects}
\usage{
\method{summary}{Npx}(object, ...)
}
\arguments{
\item{object}{An NPX object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary method for NPX objects
}
