% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{summary.segment}
\alias{summary.segment}
\title{Summary method for segment objects}
\usage{
\method{summary}{segment}(object, ...)
}
\arguments{
\item{object}{A segment object}

\item{...}{Additional arguments passed to summary}
}
\description{
Summary method for segment objects
}
