% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{track_harmonic}
\alias{track_harmonic}
\title{Track Harmonics in Audio (from warbleR)}
\usage{
track_harmonic(
  wave,
  f,
  wl = 512,
  wn = "hanning",
  ovlp = 0,
  fftw = FALSE,
  at = NULL,
  tlim = NULL,
  threshold = 10,
  bandpass = NULL,
  clip = NULL,
  plot = TRUE,
  xlab = "Times (s)",
  ylab = "Frequency (kHz)",
  ylim = c(0, f/2000),
  adjust.wl = FALSE,
  dfrq = FALSE,
  ...
)
}
\description{
Track Harmonics in Audio (from warbleR)
}
\keyword{internal}
