% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npx.R
\name{validate_npx}
\alias{validate_npx}
\title{Validate NPX Object}
\usage{
validate_npx(x)
}
\arguments{
\item{x}{Object to validate}
}
\value{
TRUE if valid, otherwise throws error
}
\description{
Validate NPX Object
}
\keyword{internal}
