% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASAP.R
\name{write_ASAP}
\alias{write_ASAP}
\title{write_ASAP}
\usage{
write_ASAP(ASAP_input, output_name)
}
\arguments{
\item{ASAP_input}{R list returned by ASAP() function}

\item{output_name}{string containing the file output name}
}
\description{
write_ASAP allows to save ASAP results in a table-like format.
}
\examples{
asap_results <- ASAP(pca_input = pca, as_file = example_as)
write_ASAP(ASAP_input = asap_results, output_name = 'my_dir/my_asap_results.txt')
}
