% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{as_segment}
\alias{as_segment}
\title{Convert data to a segment object}
\usage{
as_segment(x)
}
\arguments{
\item{x}{A data frame to convert to a segment object}
}
\value{
A validated segment object
}
\description{
Convert data to a segment object
}
\examples{
\dontrun{
df <- data.frame(
  filename = "song.wav",
  day_post_hatch = 1,
  label = "a",
  start_time = 0,
  end_time = 1,
  duration = 1
)
segment <- as_segment(df)
}

}
