% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASAP-package.R
\docType{package}
\name{ASAP-package}
\alias{ASAP}
\alias{ASAP-package}
\title{ASAP: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
