% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{create_template}
\alias{create_template}
\title{Create Correlation Templates for Song Analysis}
\usage{
create_template(x, ...)
}
\arguments{
\item{x}{An object to process, either a character path to a WAV file
or a SAP object}

\item{...}{Additional arguments passed to specific methods}
}
\value{
A correlation template object or updated SAP object
}
\description{
A generic function to create correlation templates from WAV files or SAP objects
for song detection and analysis.
}
\details{
This generic function supports template creation through two methods:
\itemize{
\item Default method for creating templates from individual WAV files
\item SAP object method for creating templates from audio clips within a SAP object
}
}
\examples{
\dontrun{
# Create template from WAV file
template <- create_template("path/to/song.wav",
                           template_name = "template1",
                           start_time = 1.0,
                           end_time = 2.0)

# Create template from SAP object
sap_obj <- create_template(sap_object,
                          template_name = "template1",
                          clip_name = "clip1")
}

}
