% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{detect_template}
\alias{detect_template}
\title{Detect Templates in Song Data}
\usage{
detect_template(x, ...)
}
\arguments{
\item{x}{An object to process, either a character path to a WAV file
or a SAP object}

\item{...}{Additional arguments passed to specific methods}
}
\value{
Detection results or updated SAP object
}
\description{
A generic function to perform template matching on audio files using
correlation-based detection.
}
\details{
This generic function supports template detection through two methods:
\itemize{
\item Default method for detecting templates in individual WAV files
\item SAP object method for batch processing multiple files
}
}
\examples{
\dontrun{
# Detect template in single WAV file
detections <- detect_template("path/to/song.wav",
                             template = template_obj)

# Detect template in SAP object
sap_obj <- detect_template(sap_object,
                          template_name = "template1",
                          day = c(30, 40))
}

}
