% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{find_motif}
\alias{find_motif}
\title{Find Motifs in Song Data}
\usage{
find_motif(x, ...)
}
\arguments{
\item{x}{An object to process, either a data frame with detection data
or a SAP object}

\item{...}{Additional arguments passed to specific methods}
}
\value{
A data frame of motif information or updated SAP object
}
\description{
A generic function to identify and extract motifs from song recordings
based on detection times.
}
\details{
This generic function supports motif extraction through two methods:
\itemize{
\item Default method for processing detection data from a data frame
\item SAP object method for processing template-based detections
}
}
\examples{
\dontrun{
# Find motifs from detection data frame
motifs <- find_motif(detections,
                     pre_time = 0.1,
                     lag_time = 0.2,
                     wav_dir = "path/to/wavs")

# Find motifs in SAP object
sap_obj <- find_motif(sap_object,
                      template_name = "template1",
                      pre_time = 0.7,
                      lag_time = 0.5)
}

}
