% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{plot_PC}
\alias{plot_PC}
\alias{plot_PC.default}
\alias{plot_PC.Sap}
\title{Plot PCA coordinates of feature embeddings}
\usage{
plot_PC(x, ...)

\method{plot_PC}{default}(x, dims = c("PC1", "PC2"), ...)

\method{plot_PC}{Sap}(
  x,
  segment_type = c("motifs", "syllables", "bouts", "segments"),
  dims = c("PC1", "PC2"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{A data frame or a SAP object.}

\item{...}{Additional arguments passed to \code{\link{plot_umap}}.}

\item{dims}{Character vector of length 2 specifying the PCA dimensions to plot. Default is \code{c("PC1", "PC2")}.}

\item{segment_type}{For SAP objects: Type of segments to visualize ('motifs', 'syllables', 'bouts', 'segments')}

\item{verbose}{For SAP objects: Whether to print progress messages}
}
\description{
Wrapper for \code{\link{plot_umap}} to plot PCA coordinates (PC1 and PC2) instead of UMAP.
Inherits all arguments from \code{\link{plot_umap}}.
}
