% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_PCA.R
\name{read_flash}
\alias{read_flash}
\title{read_flash}
\usage{
read_flash(pca_input)
}
\arguments{
\item{pca_input}{point to the directory and flash pca file}
}
\value{
Returns a PCA matrix with new header: POP ID PC1 PCN
}
\description{
read_flash reads PCA output from flashpca, where the first column contains Family ID, the second sample ID, and the following the PCs
}
\examples{
\dontrun{
read_flash(pca_input = 'data/TOY_flash.pca')
}
}
