% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{run_pca}
\alias{run_pca}
\title{Run Principal Component Analysis}
\usage{
run_pca(x, ...)
}
\arguments{
\item{x}{An object to analyze, either a matrix/data frame
or a SAP object}

\item{...}{Additional arguments passed to specific methods}
}
\value{
PCA results or updated SAP object
}
\description{
A generic function to perform PCA with support for multiple methods
and large-scale data processing.
}
\details{
This generic function supports PCA through two methods:
\itemize{
\item Default method for matrices with multiple PCA implementations
\item SAP object method for organized trajectory data
}
}
\examples{
\dontrun{
# Run PCA on matrix
pca <- run_pca(matrix, method = "irlba", n_components = 50)

# Run PCA on SAP object
sap_obj <- run_pca(sap_object,
                   segment_type = "motifs",
                   data_type = "traj_mat")
}

}
