% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asap.R
\name{write_asap}
\alias{write_asap}
\title{write_asap}
\usage{
write_asap(asap_input, output_name)
}
\arguments{
\item{asap_input}{R list returned by asap() function}

\item{output_name}{string containing the file output name}
}
\description{
write_asap allows to save ASAP results in a table-like format.
}
\examples{
asap_results <- asap(pca_input = pca, as_file = example_as)
write_asap(asap_input = asap_results, output_name = 'my_dir/my_asap_results.txt')
}
